DummyEmployeeDataRepository - combine EmployeeDataRepositoryImpl and DummyEmployeeDataUtils
